requirements.dev.txt
asgiref==3.2.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==3.0
django-cors-headers==3.2.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
ndg-httpsclient==0.5.1
pusher==2.1.4
pyasn1==0.4.8
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
